backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.6.2
attrs==25.1.0
autokey==0.96.0
Beaker==1.12.1
beautifulsoup4==4.13.4
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.38.46
botocore==1.38.46
Brlapi==0.8.6
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-341-build/cockpit-341/tmp/wheel/cockpit-341-py3-none-any.whl#sha256=1dd0889753d4629153a62ed95f46f8fc113ede8579034a75d0097cf7fd002295
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnf==4.23.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-slim==0.115.12
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.58.5
fpdf2==2.7.9
fros==1.1
Glances==4.3.0.8
greenlet==3.2.3
gunicorn==23.0.0
h11==0.14.0
humanize==4.12.0
idna==3.10
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.68
Levenshtein==0.26.1
libcomps==0.1.21
libdnf==0.74.0
libvirt-python==11.0.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
mercurial==6.9.5
nftables==0.1
numpy==2.2.6
olefile==0.47
orjson==3.10.18
packaging==24.2
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.3.8
ply==3.11
productmd==1.45
psutil==6.1.1
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.62
pyOpenSSL==25.0.0
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-xlib==0.33
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.15.4-build/kernel-6.15.4/linux-6.15.4-200.fc42.x86_64/tools/net/ynl
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.25.0
rpm==4.20.1
s3transfer==0.13.0
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.9.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.0
systemd-python==235
trash-cli==0.24.5.26
typing_extensions==4.12.2
ulauncher==5.15.7
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.31.2
websocket-client==1.8.0
Werkzeug==3.1.3
xkbregistry==0.3
